wopsie.. forgot to commit.. big change though

+++ M2M-assignment/dht.h
@@ -14,7 +14,7 @@ public:
   float getFarenheit();
   float getHumidity();
 private:
-  int number_of_readings;
+  float number_of_readings;
 };
